Add validation to volume attach/detach/delete sub-commands
They can only manage volume if it's not linked to Kubernetes cluster

+++ cmd/volume_remove.go
@@ -47,6 +47,17 @@ var volumeRemoveCmd = &cobra.Command{
 			}
 		}
 
+		if !utility.CanManageVolume(volume) {
+			cluster, err := client.FindKubernetesCluster(volume.ClusterID)
+			if err != nil {
+				utility.Error("Unable to find cluster - %s", err)
+				os.Exit(1)
+			}
+
+			utility.Error("Unable to %s this volume because it's being managed by your %q Kubernetes cluster", cmd.Name(), cluster.Name)
+			os.Exit(1)
+		}
+
 		if utility.UserConfirmedDeletion("volume", defaultYes, volume.Name) {
 
 			_, err = client.DeleteVolume(volume.ID)
